Alternate alias file implementation, commit 2 of 2
See 8b0934b0adcc57e7c62803ee07f03cde2a115298 for first commit, it's split in two because otherwise GitHub mixes up the changes in a way that is hard to understand.

So, I was having trouble with this plugin, and noticed it would generate the files and then delete them before the task was finished. I eventually rewrote the plugin to write the file at the end, rather than the beginning, as shown here.

But, UGH, I discovered that the reason it was deleting my file was because I had `rake preview` running in another terminal tab, which hadn't been restarted to use this plugin, and thus didn't recognize the extra files and was deleting them after the `rake generate` with this plugin wrote them. Sigh, I can't believe I did that, but lesson learned about Jekyll, I guess.